refactor(ui): generic useBadgeCount + shared link class const (spec-379)
spec-355 audit dry-11 + dry-12, behaviour-preserving.

dry-11: collapse the twin nav-badge hooks useDriftInboxCount /
useMyIssuesCount into one generic useBadgeCount(fetchItems, enabled).
The single axis of difference is the bound fetch thunk; fetchItems is
held in a ref so reload's deps stay [enabled, pathname] — reload
identity stays stable across renders, reproducing both originals'
trigger set (mount / enabled / pathname / SSE) exactly. The two named
hooks become thin wrappers passing module-stable thunks; public
signatures and AppShell call sites unchanged.

dry-12: hoist the byte-identical inline-cite button className template
shared by DecisionLink and TaskLink into one module-level
linkButtonClassName(error) helper; both buttons use it, yielding
byte-identical class strings for the error and default branches.

Lean verify: tsc -b clean, targeted vitest green (useDriftInboxCount,
DecisionLink, AppShell — 43 passed), biome clean over touched files.
